pyproject.toml
[project]
name = "thunderbird-assignment"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "ipykernel>=6.30.1",
    "langgraph>=0.2.0",
    "langchain>=0.3.0",
    "langchain-openai>=0.2.0",
    "langchain-anthropic>=0.2.0",
    "pydantic>=2.0.0",
    "openai>=1.0.0",
    "anthropic>=0.34.0",
    "pandas>=2.0.0",
    "jinja2>=3.1.0",
    "rich>=13.0.0",
    "typer>=0.12.0",
    "python-dotenv>=1.0.0",
    "matplotlib>=3.7.0",
    "graphviz>=0.20.0",
    "plotly>=5.17.0",
    "kaleido>=0.2.1",
    "flask>=3.0.0",
]
